[♻️refactor]: convertLanguageValueToKey 함수 분리
