+++ src/pages/HomePage/HomePage.tsx
@@ -2,8 +2,8 @@ import { Spinner } from '@/components';
 import { useRoomsList } from '@/hooks/Api/useRooms';
 import useFilterStore from '@/store/RoomsListStore/useFilterStore';
 import usePageStore from '@/store/RoomsListStore/usePageStore';
-import { LanguagesType } from '@/types/room';
 
+import { convertLanguageValueToKey } from './convertValueTokey';
 import HomeFooter from './HomeFooter/HomeFooter';
 import HomeNav from './HomeNav/HomeNav';
 import * as S from './HomePage.style';
@@ -16,18 +16,6 @@ export default function HomePage() {
   const { searchTitle, selectedLanguage, selectedAccess, selectedStatus } =
     useFilterStore();
 
-  const convertLanguageValueToKey = (value: string[]) => {
-    const languageKey: { [key: string]: LanguagesType } = {
-      Java: 'JAVA',
-      파이썬: 'PYTHON',
-      자바스크립트: 'JAVASCRIPT',
-      'C++': 'C++',
-    };
-
-    return value.map(v => languageKey[v]);
-  };
-
-  // Todo: 언어 제대로 필터 안되는 현상 수정
   const languages = convertLanguageValueToKey(selectedLanguage).join(', ');
   const roomAccessType = selectedAccess === 'PRIVATE' ? 'PRIVATE' : undefined;
   const roomStatus = selectedStatus === 'RECRUITING' ? 'RECRUITING' : undefined;
@@ -48,7 +36,6 @@ export default function HomePage() {
   return (
     <S.HomePageContainer>
       <S.HomePageWrapper>
-        {/* // Todo: refetch 함수 넘기는 방법 */}
         {/* 상단 Nav */}
         <HomeNav refetch={refetch} />
